requirements.txt
openai==1.78.1
PyPDF2==3.0.1
python-dotenv==1.1.0
streamlit==1.45.0
